pyproject.toml
[project]
name = "p4nd0-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "discord-py>=2.7.1",
    "mysql-connector-python>=9.6.0",
    "pytest==9.0.2",
    "python-dotenv>=1.2.2",
    "requests>=2.32.5",
]
